requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-Login==0.6.2
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.3
pandas==2.0.2
profanity==1.1
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
rake-nltk==1.0.6
regex==2023.5.5
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
zipp==3.15.0
